+++ fabric-data-generation-api-v1/src/main/java/net/fabricmc/fabric/api/datagen/v1/FabricModel.java
@@ -18,9 +18,9 @@ package net.fabricmc.fabric.api.datagen.v1;
 
 import net.minecraft.data.client.Model;
 
-import net.fabricmc.fabric.api.datagen.v1.builder.DisplayBuilder;
-import net.fabricmc.fabric.api.datagen.v1.builder.ElementBuilder;
-import net.fabricmc.fabric.api.datagen.v1.builder.OverrideBuilder;
+import net.fabricmc.fabric.api.datagen.v1.model.DisplayBuilder;
+import net.fabricmc.fabric.api.datagen.v1.model.ElementBuilder;
+import net.fabricmc.fabric.api.datagen.v1.model.OverrideBuilder;
 
 /**
  * Fabric-provided extensions for {@link Model}.
