+++ fabric-data-generation-api-v1/src/main/java/net/fabricmc/fabric/api/datagen/v1/model/FaceBuilder.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.api.datagen.v1.builder;
+package net.fabricmc.fabric.api.datagen.v1.model;
 
 import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
